feat/SW-38/refactor(Back): PostgreSQL 마이그레이션에 따른 마이페이지 및 AOP 수정
* Note: 추후 고도화 예정

+++ src/main/java/com/web/SearchWeb/aop/OwnerCheckAspect.java
@@ -18,6 +18,11 @@ import org.springframework.stereotype.Component;
 import java.util.Arrays;
 import java.util.Objects;
 
+/**
+ * OwnerCheckAspect
+ * 
+ * 리소스 소유자 검증 AOP
+ */
 @Aspect
 @Slf4j(topic = "[OwnerCheckAspect]")
 @Component
@@ -34,14 +39,14 @@ public class OwnerCheckAspect {
     @Before("@annotation(ownerCheck)")
     public void validateOwner(JoinPoint joinPoint, OwnerCheck ownerCheck) {
         // 접근 검증 대상 리소스의 ID 추출
-        Integer targetId = extractTargetIdFromParams(joinPoint, ownerCheck.idParam());
+        Long targetId = extractTargetIdFromParams(joinPoint, ownerCheck.idParam());
 
         // 현재 로그인한 사용자의 memberId 추출
         Authentication auth = validateAuthenticatedUser();
-        Integer currentUserId = extractMemberId(auth);
+        Long currentUserId = extractMemberId(auth);
 
         // 서비스 이름에 따라 리소스 작성자 memberId 조회
-        Integer ownerId = findOwnerIdByServiceName(ownerCheck.service(), targetId);
+        Long ownerId = findOwnerIdByServiceName(ownerCheck.service(), targetId);
 
         // 현재 사용자와 리소스 소유자 검증
         if (!Objects.equals(currentUserId, ownerId)) {
@@ -54,21 +59,22 @@ public class OwnerCheckAspect {
 
 
     /**
-     * 접근 검증 대상이 되는 리소스의 ID를 파라미터 이름(idParam)을 통해 찾아 Integer로 반환
+     * 접근 검증 대상이 되는 리소스의 ID를 파라미터 이름(idParam)을 통해 찾아 Long으로 반환
      *  ex) @OwnerCheck(idParam = "boardId", ...) -> 메서드의 boardId 값을 찾아 사용
      *
      * @param joinPoint 현재 실행된 메서드의 실행 정보
      * @param idParam   검증 대상 리소스 ID의 파라미터 이름 (예: "boardId" 문자열)
      * @return 접근 검증 대상이 되는 리소스의 ID 값
      */
-    private Integer extractTargetIdFromParams(JoinPoint joinPoint, String idParam) {
+    private Long extractTargetIdFromParams(JoinPoint joinPoint, String idParam) {
         Object[] args = joinPoint.getArgs(); // 메서드 실제 인자 값 배열
         MethodSignature signature = (MethodSignature) joinPoint.getSignature();
         String[] paramNames = signature.getParameterNames(); // 메서드 파라미터 이름 배열
 
         for (int i = 0; i < paramNames.length; i++) {
             if (paramNames[i].equals(idParam)) {
-                return Integer.parseInt(args[i].toString());
+                // Integer나 Long 모두 지원하도록 String으로 변환 후 parse
+                return Long.parseLong(args[i].toString());
             }
         }
         log.error("{}' 파라미터를 찾을 수 없음. 실제 파라미터: {}", idParam, Arrays.toString(paramNames));
@@ -76,7 +82,7 @@ public class OwnerCheckAspect {
     }
 
 
-    // SecurityContext 에서 인증된 사용자 반한
+    // SecurityContext 에서 인증된 사용자 반환
     private Authentication validateAuthenticatedUser() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         if (auth == null || !auth.isAuthenticated() || "anonymousUser".equals(auth.getPrincipal())) {
@@ -87,8 +93,8 @@ public class OwnerCheckAspect {
     }
 
 
-    // 인증 객체에서 현재 로그인한 사용자 memberId 추출
-    private Integer extractMemberId(Authentication auth) {
+    // 인증 객체에서 현재 로그인한 사용자 memberId 추출 (Long 반환)
+    private Long extractMemberId(Authentication auth) {
         Object principal = auth.getPrincipal();
         if (principal instanceof CustomUserDetails u) return u.getMemberId();
         if (principal instanceof CustomOAuth2User u) return u.getMemberId();
@@ -98,11 +104,11 @@ public class OwnerCheckAspect {
 
 
     // 서비스 이름에 따라 해당 리소스의 작성자 조회
-    private Integer findOwnerIdByServiceName(String service, Integer targetId) {
+    private Long findOwnerIdByServiceName(String service, Long targetId) {
         return switch (service) {
             case "boardService"  -> boardService.findMemberIdByBoardId(targetId);
             case "commentService" -> commentService.findMemberIdByCommentId(targetId);
-            case "memberService" -> targetId;
+            case "memberService" -> targetId; // memberService의 경우 targetId가 곧 memberId
             default -> {
                 log.error("지원하지 않는 서비스명 '{}'", service);
                 throw new IllegalArgumentException("지원하지 않는 서비스명입니다.");
